feat(ui): Enhance agent list with card layout and timestamps
Add styled agent cards with improved visual hierarchy and information:
- Implement grid layout with gradient-styled cards
- Add last updated timestamp with relative formatting
- Move formatTimestamp utility to shared utils
- Add DataList component for consistent data presentation
- Improve layout spacing and visual hierarchy

This change enhances the agent list view with a more polished and
informative presentation while making the timestamp formatting utility
more reusable across the application.
